Небольшое рефакторинг. Перенес даты и функции
